Rename classes Restructure folders
