+++ docker/Dockerfile_coinbase_producer
@@ -16,4 +16,4 @@ ENV PYTHONPATH=/app/src
 ENV PYTHONUNBUFFERED=1
 ENV IN_DOCKER=1
 # Start de app
-CMD ["python", "src/coinbase_producer/main.py"]
+CMD ["python", "src/producers/coinbase_producer/main.py"]
