+++ src/interfaces/BaseStreamProducer.py
